feat: introduce new DTO and entity interfaces for improved data handling and security refactor: enhance Http utility with not found handling method chore: remove deprecated security classes and interfaces
