+++ src/main/kotlin/ru/virgil/spring/tools/dto/IdentifiedDto.kt
@@ -1,0 +1,8 @@
+package ru.virgil.spring.tools.dto
+
+import java.util.*
+
+interface IdentifiedDto {
+
+    var uuid: UUID?
+}
