+++ src/main/kotlin/ru/virgil/spring/tools/dto/TimedDto.kt
@@ -1,0 +1,9 @@
+package ru.virgil.spring.tools.dto
+
+import java.time.ZonedDateTime
+
+interface TimedDto {
+
+    var createdAt: ZonedDateTime?
+    var updatedAt: ZonedDateTime?
+}
