+++ src/main/kotlin/ru/virgil/spring/tools/image/PrivateImageInterface.kt
@@ -1,7 +1,7 @@
 package ru.virgil.spring.tools.image
 
-import ru.virgil.spring.tools.util.data.Identified
-import ru.virgil.spring.tools.util.data.Owned
+import ru.virgil.spring.tools.entity.Identified
+import ru.virgil.spring.tools.entity.Owned
 import java.nio.file.Path
 
 interface PrivateImageInterface : Owned, Identified {
