--- src/main/kotlin/ru/virgil/spring/tools/security/oauth/SecurityUser.kt
@@ -1,9 +1,0 @@
-package ru.virgil.spring.tools.security.oauth
-
-import ru.virgil.spring.tools.Deprecation
-
-@Deprecated(Deprecation.NEW_NATIVE_AUTH)
-interface SecurityUser : UserDetailsKt {
-
-    val firebaseUserId: String
-}
