+++ src/main/kotlin/ru/virgil/spring/tools/security/user/SimpleCreatorAuditorAware.kt
@@ -1,7 +1,8 @@
-package ru.virgil.spring.tools.security
+package ru.virgil.spring.tools.security.user
 
 import org.springframework.data.domain.AuditorAware
 import org.springframework.stereotype.Component
+import ru.virgil.spring.tools.security.Security
 import java.util.*
 
 @Component
