+++ src/main/kotlin/ru/virgil/spring/tools/security/user/UserRoles.kt
@@ -1,8 +1,8 @@
-package ru.virgil.spring.tools.security
+package ru.virgil.spring.tools.security.user
 
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 
-object Roles {
+object UserRoles {
 
     fun Enum<*>.toAuthority(): SimpleGrantedAuthority {
         return SimpleGrantedAuthority(this.name)
