--- src/main/kotlin/ru/virgil/spring/tools/util/data/Created.kt
@@ -1,10 +1,0 @@
-package ru.virgil.spring.tools.util.data
-
-import ru.virgil.spring.tools.Deprecation
-import java.time.ZonedDateTime
-
-@Deprecated(Deprecation.REDUNDANT)
-interface Created {
-
-    val createdAt: ZonedDateTime
-}
