--- src/main/kotlin/ru/virgil/spring/tools/util/data/Updated.kt
@@ -1,10 +1,0 @@
-package ru.virgil.spring.tools.util.data
-
-import ru.virgil.spring.tools.Deprecation
-import java.time.ZonedDateTime
-
-@Deprecated(Deprecation.REDUNDANT)
-interface Updated {
-
-    val updatedAt: ZonedDateTime
-}
